% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_mask.R
\name{canopy_mask}
\alias{canopy_mask}
\title{Create canopy mask}
\usage{
canopy_mask(
  seg_class_rts,
  out_rts,
  canopy_classes,
  canopy_edits = NULL,
  openings = 1,
  opening_radius = 0.5,
  ...
)
}
\arguments{
\item{canopy_classes}{integer. Raster values for \code{seg_class_rts} that correspond to canopy}

\item{openings}{the number of times that a morphological opening will be applied}

\item{opening_radius}{radius of morphological opening window}
}
\description{
Create canopy mask
}
