% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Canopy.R
\name{CanopyMask}
\alias{CanopyMask}
\title{Create canopy mask}
\usage{
CanopyMask(
  segClassRas_RSDS,
  canopyMask_RSDS,
  canopyClasses,
  openings = 1,
  openingRadius = 0.5,
  tileNames = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{canopyClasses}{names of raster classes that are included in the canopy}

\item{openings}{the number of times that a morphological opening will be applied}

\item{openingRadius}{radius of morphological opening window}
}
\description{
Create canopy mask
}
