% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classifier_create}
\alias{classifier_create}
\title{Create classifier}
\usage{
classifier_create(
  training_pts,
  classifier_file = NULL,
  seg_id,
  predictors = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\description{
Create classifier
}
