% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_seg_ras}
\alias{classify_seg_ras}
\title{Classify raster segments}
\usage{
classify_seg_ras(
  seg_class_poly_rsds,
  seg_ras_rsds,
  seg_class_ras_rsds,
  segClasses,
  seg_id,
  tile_names = NULL,
  overwrite = FALSE
)
}
\description{
This needs \code{classify_seg_poly} to be run first
}
