% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Classification.R
\name{ClassifyRaster}
\alias{ClassifyRaster}
\title{Classify raster segments}
\usage{
ClassifyRaster(
  segClassPoly_RSDS,
  segRas_RSDS,
  segClassRas_RSDS,
  segClasses,
  segID,
  tileNames = NULL,
  clusters = 1,
  overwrite = FALSE
)
}
\description{
Classify raster segments
}
