% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Classification.R
\name{ClassifySegments}
\alias{ClassifySegments}
\title{Classify segments}
\usage{
ClassifySegments(
  classifierFile,
  segPoly_RSDS,
  segClassPoly_RSDS,
  classEdits,
  metrics,
  segID,
  tileNames = NULL,
  overwrite = FALSE
)
}
\description{
Classify segments
}
