% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Classification.R
\name{CreateClassifier}
\alias{CreateClassifier}
\title{Create classifier}
\usage{
CreateClassifier(
  trainingPts,
  classifierFile = NULL,
  segID,
  predictors = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\description{
Create classifier
}
