% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.exe_tile_worker}
\alias{.exe_tile_worker}
\title{Run a worker in parallel or serial}
\usage{
.exe_tile_worker(
  tile_names,
  worker,
  cluster_vts = NULL,
  clusters = getOption("misterRS.clusters")
)
}
\description{
Run a worker in parallel or serial
}
