% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{.lockSave}
\alias{.lockSave}
\title{Attempt to save a SHP file that may have a lock}
\usage{
.lockSave(output, outPath)
}
\description{
Certain programs (such as QGIS) may cause a file to be corrupted when an attempt is made to
overwrite it while it's locked. If this happens, this function will save a backup file
}
