% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Classification.R
\name{ExtractTrainingData}
\alias{ExtractTrainingData}
\title{Extract Training data}
\usage{
ExtractTrainingData(
  trainingPts,
  segPoly_RSDS,
  metrics,
  segID,
  overwrite = FALSE
)
}
\description{
Extract Training data
}
