% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{final_canopy}
\alias{final_canopy}
\title{Final Canopy}
\usage{
final_canopy(
  trees_class_ras_rsds,
  canopyClasses,
  boundary,
  out_file,
  boundary_buff = 0,
  tile_names = NULL
)
}
\description{
Applies the following process:
\enumerate{
\item Clip to boundary
\item Remove any 'NotTree' classes from canopy
\item Merge into final TIF file
}
}
