% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Final.R
\name{FinalCanopy}
\alias{FinalCanopy}
\title{1. Clip to boundary
2. Remove any 'NotTree' classes from canopy
3. Merge into final TIF file}
\usage{
FinalCanopy(
  treeClassRas_RSDS,
  canopyClasses,
  boundary,
  out_file,
  boundary_buff = 0,
  tileNames = NULL
)
}
\description{
1. Clip to boundary
2. Remove any 'NotTree' classes from canopy
3. Merge into final TIF file
}
