% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Surfaces.R
\name{MakeDSM}
\alias{MakeDSM}
\title{Make Digital Surface Model}
\usage{
MakeDSM(
  in_cat,
  DEM_RSDS = NULL,
  out_RSDS,
  res = 0.25,
  zMin = 0,
  zMax = 80,
  max_edge = c(0, 1),
  subcircle = 0,
  thresholds = c(0, 2, 5, 10, 15, 20, 25, 30, 35, 40),
  LAS_select = "xyzcr",
  LAS_classes = NULL,
  tileNames = NULL,
  overwrite = FALSE
)
}
\description{
Generate a DSM. If \code{DEM_RSDS} is provided, it will generate a normalized DSM (nDSM)
}
