% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Surfaces.R
\name{MakeNDSM}
\alias{MakeNDSM}
\title{Make Normalized Digital Surface Model}
\usage{
MakeNDSM(
  in_cat,
  DEM_RSDS,
  out_RSDS,
  nDSMres,
  zMin,
  zMax,
  maxEdge = c(0, 1),
  subCircle = 0,
  thresholds = c(0, 2, 5, 10, 15, 20, 25, 30, 35, 40),
  LASselect = "xyzcr",
  LASclasses = NULL,
  tileNames = NULL,
  overwrite = FALSE
)
}
\description{
This version will create an nDSM directly from the point cloud (no existing DSM required)
}
