% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Surfaces.R
\name{MakeNDSM2}
\alias{MakeNDSM2}
\title{Make Normalized Digital Surface Model}
\usage{
MakeNDSM2(DSM_RSDS, DEM_RSDS, out_RSDS, tileNames = NULL, overwrite = FALSE)
}
\description{
This version will subtract the DEM from an existing DSM
}
