% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\title{Mask}
\usage{
mask(
  in_rsds,
  out_rsds,
  mask_rsds,
  mask_na = FALSE,
  tile_names = NULL,
  overwrite = FALSE,
  use_neibs = FALSE
)
}
\arguments{
\item{use_neibs}{logical. Set to TRUE if you wish to use include buffered sections from neighbouring tiles.}
}
\description{
Mask
}
