% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask_rs}
\alias{mask_rs}
\title{Mask}
\usage{
mask_rs(in_rts, out_rts, mask_rts, mask_na = FALSE, use_neibs = FALSE, ...)
}
\arguments{
\item{use_neibs}{logical. Set to TRUE if you wish to use include buffered sections from neighbouring tiles.}
}
\description{
Mask
}
