% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\title{Merge multiple RSDS into a single one according to zones}
\usage{
merge(
  in_rsds,
  out_rsds,
  zones_path,
  zone_attr,
  tile_names = NULL,
  overwrite = FALSE
)
}
\description{
Merge multiple RSDS into a single one according to zones
}
