% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_rs}
\alias{merge_rs}
\title{Merge multiple RSDS into a single one according to zones}
\usage{
merge_rs(in_rsds, out_rsds, zones_path, zone_field, id_field = NULL, ...)
}
\description{
Merge multiple RSDS into a single one according to zones
}
