% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaic_rs}
\alias{mosaic_rs}
\title{Mosaic a RS Dataset}
\usage{
mosaic_rs(
  rsds,
  overlap = "nbuffs",
  shp_clip = FALSE,
  outFile = NULL,
  overwrite = FALSE,
  tile_names = NULL
)
}
\description{
Mosaic a RS Dataset
}
