% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Segments.R
\name{MSS}
\alias{MSS}
\title{Mean Segment Shift (deprecated)}
\usage{
MSS(
  image_RSDS,
  segPoly_RSDS,
  segRas_RSDS,
  spat = 19.5,
  spec = 17,
  mins = 40,
  segID = "polyID",
  tileNames = NULL,
  clusters = 1,
  overwrite = FALSE
)
}
\description{
WARNING: As of 2019-12-11, this function is semi-deprecated.
This version of the MSS algorithm processes each tile individually.
It appears preferable to generate all segments at once AND THEN chop the
output into individual tiles, for which the \code{MSS2} and \code{MSS2chop}
have been made available.
}
