% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Segments.R
\name{MSS2chop}
\alias{MSS2chop}
\title{Mean Segment Shift 2 - Chop GeoPackage Into Tiles}
\usage{
MSS2chop(in_GPKG, segPoly_RSDS, chunk_size = 2000, segID = "polyID")
}
\description{
After having run \code{MSS2}, use this function to chop the output
into tiles while managing polygons that cross into the tiles' buffers effectively
}
