% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_PCA.R
\name{PCA}
\alias{PCA}
\title{Principal Component Analysis}
\usage{
PCA(
  ortho_RSDS,
  out_RSDS,
  nComp = 2,
  tileNames = NULL,
  clusters = 1,
  overwrite = FALSE
)
}
\description{
Principal Component Analysis
}
