% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_global}
\alias{pca_global}
\title{Principal Component Analysis (Global)}
\usage{
pca_global(img_rts, out_rts, pca_model, n_comp = 2, in_bands = c(1, 2, 3), ...)
}
\description{
Second approach: create a SINGLE PCA (using \code{pca_model()}) and then apply it to each tile.
This looks better from an ensemble view, but may not accentuate variation as much as the
other approach.
}
