% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_local}
\alias{pca_local}
\title{Principal Component Analysis - local}
\usage{
pca_local(
  img_rsds,
  out_rsds,
  n_comp = 2,
  in_bands = c(1, 2, 3),
  spca = TRUE,
  ...
)
}
\description{
This is the first approach I tried: calculate a different PCA for each tile.
This means that the tiles don't necessarily match very well (edge effects are present),
but it does maximize local variation.
}
