% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_model}
\alias{pca_model}
\title{Principal Component Analysis model}
\usage{
pca_model(
  img_rts,
  out_file,
  nSamples = NULL,
  in_bands = c(1, 2, 3),
  removeBlack = T
)
}
\description{
Create a single PCA model for the entire area. Used in conjunction with \code{pca_global}
}
