% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_PCA.R
\name{PCAglobal}
\alias{PCAglobal}
\title{Principal Component Analysis - global}
\usage{
PCAglobal(
  ortho_RSDS,
  out_RSDS,
  PCA_model,
  nComp = 2,
  in_bands = c(1, 2, 3),
  tileNames = NULL,
  overwrite = FALSE
)
}
\description{
Second approach: create a SINGLE PCA (using `PCAmodel`) and then apply it to each tile.
This looks better from an ensemble view, but may not accentuate variation as much as the
other approach.
}
