% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_PCA.R
\name{PCAlocal}
\alias{PCAlocal}
\title{Principal Component Analysis - local}
\usage{
PCAlocal(
  ortho_RSDS,
  out_RSDS,
  nComp = 2,
  in_bands = c(1, 2, 3),
  tileNames = NULL,
  overwrite = FALSE
)
}
\description{
This is the first approach I tried: calculate a different PCA for each tile.
This means that the tiles don't necessarily match very well (edge effects are present),
but it does maximize local variation.
}
