% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_PCA.R
\name{PCAmodel}
\alias{PCAmodel}
\title{Principal Component Analysis model}
\usage{
PCAmodel(
  ortho_RSDS,
  out_file,
  nSamples = NULL,
  in_bands = c(1, 2, 3),
  removeBlack = T,
  spca = FALSE
)
}
\description{
Create a single PCA model for the entire area
}
