% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{poly_to_ras}
\alias{poly_to_ras}
\title{Rasterize polygonal segments}
\usage{
poly_to_ras(
  seg_poly_rsds,
  seg_ras_rsds,
  res,
  seg_id = "polyID",
  tile_names = NULL,
  overwrite = FALSE
)
}
\description{
Rasterize polygonal segments
}
