% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extents.R
\name{qgis_frame_to_ext}
\alias{qgis_frame_to_ext}
\title{Convert frame copy/pasted from QGIS into an extent}
\usage{
qgis_frame_to_ext(frame)
}
\description{
Convert frame copy/pasted from QGIS into an extent
}
