% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Segments.R
\name{RasterSegment}
\alias{RasterSegment}
\title{Rasterize polygonal segments}
\usage{
RasterSegment(
  segPoly_RSDS,
  segRas_RSDS,
  res,
  segID = "polyID",
  tileNames = NULL,
  clusters = 1,
  overwrite = FALSE
)
}
\description{
Rasterize polygonal segments
}
