% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Rescale.R
\name{Rescale}
\alias{Rescale}
\title{Rescale raster dataset}
\usage{
Rescale(
  in_RSDS,
  out_RSDS,
  from,
  to = c(0, 254),
  NAtoZero = TRUE,
  dataType = "INT1U",
  tileNames = NULL,
  clusters = 1,
  overwrite = FALSE
)
}
\description{
Rescale raster dataset
}
