% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_misc.R
\name{Retile}
\alias{Retile}
\title{Retile dataset}
\usage{
Retile(
  inFiles,
  out_RSDS,
  res,
  bands = NULL,
  tileNames = NULL,
  clusters = 1,
  overwrite = FALSE,
  makeVRTlist = FALSE
)
}
\description{
Retile dataset
}
