% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_xts.R
\name{rts}
\alias{rts}
\title{Raster tileset}
\description{
Raster tileset.
}
\section{Super class}{
\code{\link[misterRS:xts]{misterRS::xts}} -> \code{rts}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rts-new}{\code{rts$new()}}
\item \href{#method-rts-tile_dir}{\code{rts$tile_dir()}}
\item \href{#method-rts-tile_path}{\code{rts$tile_path()}}
\item \href{#method-rts-available_tiles}{\code{rts$available_tiles()}}
\item \href{#method-rts-print}{\code{rts$print()}}
\item \href{#method-rts-has_tiles}{\code{rts$has_tiles()}}
\item \href{#method-rts-complete}{\code{rts$complete()}}
\item \href{#method-rts-clone}{\code{rts$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rts-new"></a>}}
\if{latex}{\out{\hypertarget{method-rts-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rts$new(id, name, dir, file_ext = "tif")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rts-tile_dir"></a>}}
\if{latex}{\out{\hypertarget{method-rts-tile_dir}{}}}
\subsection{Method \code{tile_dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rts$tile_dir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rts-tile_path"></a>}}
\if{latex}{\out{\hypertarget{method-rts-tile_path}{}}}
\subsection{Method \code{tile_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rts$tile_path(tile_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rts-available_tiles"></a>}}
\if{latex}{\out{\hypertarget{method-rts-available_tiles}{}}}
\subsection{Method \code{available_tiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rts$available_tiles(ts = getOption("misterRS.ts"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rts-print"></a>}}
\if{latex}{\out{\hypertarget{method-rts-print}{}}}
\subsection{Method \code{print()}}{
Print.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rts$print(ts = getOption("misterRS.ts"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rts-has_tiles"></a>}}
\if{latex}{\out{\hypertarget{method-rts-has_tiles}{}}}
\subsection{Method \code{has_tiles()}}{
Check if has tiles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rts$has_tiles(tile_names, attribute)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tile_names}}{Name of tile.}

\item{\code{attribute}}{Name of attribute.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rts-complete"></a>}}
\if{latex}{\out{\hypertarget{method-rts-complete}{}}}
\subsection{Method \code{complete()}}{
Is this VTS complete?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rts$complete(attribute, tile_names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attribute}}{Name of attribute.}

\item{\code{tile_names}}{Name of tiles. If set to NULL, all tiles will be selected.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
