% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_metrics.R
\name{seg_metrics_las}
\alias{seg_metrics_las}
\title{Segment Metrics - LAS}
\usage{
seg_metrics_las(
  seg_rts,
  seg_poly_vts,
  in_cat,
  dem_rts,
  out_rts,
  z_min,
  z_max,
  seg_id,
  prefix = NULL,
  is_ground_classified = NULL,
  is_full_classified = NULL,
  is_rgb = NULL,
  is_intensity = NULL,
  ...
)
}
\description{
Segment Metrics - LAS
}
