% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_metrics.R
\name{seg_metrics_spec}
\alias{seg_metrics_spec}
\title{Segment Metrics - Spectral}
\usage{
seg_metrics_spec(
  seg_rts,
  seg_vts,
  img_rts,
  attribute_set,
  bands = c(R = 1, G = 2, B = 3),
  zonalFun = c("mean", "sd"),
  ...
)
}
\description{
Segment Metrics - Spectral
}
