% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_metrics.R
\name{seg_metrics_tex}
\alias{seg_metrics_tex}
\title{Segment Metrics - Textural}
\usage{
seg_metrics_tex(
  seg_rts,
  seg_vts,
  img_rts,
  attribute_set,
  band = 1,
  discretize_range = NULL,
  n_grey = 16,
  ...
)
}
\description{
Segment Metrics - Textural
}
