% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segment_mss}
\alias{segment_mss}
\title{Mean Segment Shift 2 - Segment Into Single GeoPackage}
\usage{
segment_mss(
  img_rts,
  out_gpkg,
  spat = 19.5,
  spec = 17,
  mins = 40,
  writeVectoreMode = "ulu",
  seg_id = "polyID",
  min_per_tile_est = 3.5,
  ...
)
}
\description{
This variation of the MSS algorithm will write all segments into one
giant GeoPackage file. The CLI algorithm's internal tiling system is used to avoid
overloading memory, and the algorithm also does a decent job at stiching tiles together
afterwards.
}
\details{
HOWEVER: the stitching process is sometimes interrupted by invalid geometry, leaving
weird-looking breaks in some of the larger polygons

Once \code{MSS2} has been run, use \code{MSS2chop} to then re-tile the output
}
