% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segment_watershed}
\alias{segment_watershed}
\title{Watershed Segmentation}
\usage{
segment_watershed(
  out_rsds,
  chm_rsds,
  ttops_rsds,
  minCrownHgt = 0.3,
  tile_names = NULL,
  overwrite = FALSE
)
}
\description{
Watershed Segmentation
}
