% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_SegMetrics.R
\name{SegMetricsLAS}
\alias{SegMetricsLAS}
\title{Segment Metrics - LAS}
\usage{
SegMetricsLAS(
  segRas_RSDS,
  segPoly_RSDS,
  LAS_RSDS,
  DEM_RSDS,
  out_RSDS,
  metricFunc,
  zMin,
  zMax,
  segID,
  LASselect = "xyzcRGB",
  tileNames = NULL,
  overwrite = FALSE
)
}
\description{
Segment Metrics - LAS
}
