% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_SegMetrics.R
\name{SegMetricsSpec}
\alias{SegMetricsSpec}
\title{Segment Metrics - Spectral}
\usage{
SegMetricsSpec(
  segRas_RSDS,
  segPoly_RSDS,
  ortho_RSDS,
  out_RSDS,
  bands = c(R = 1, G = 2, B = 3),
  zonalFun = c("mean", "sd"),
  segID,
  tileNames = NULL,
  clusters = 1,
  overwrite = FALSE
)
}
\description{
Segment Metrics - Spectral
}
