% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_SegMetrics.R
\name{SegMetricsTex}
\alias{SegMetricsTex}
\title{Segment Metrics - Textural}
\usage{
SegMetricsTex(
  segRas_RSDS,
  segPoly_RSDS,
  img_RSDS,
  out_RSDS,
  segID,
  band = 1,
  tileNames = NULL,
  overwrite = FALSE,
  prefix = ""
)
}
\description{
Segment Metrics - Textural
}
