% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaces.R
\name{surface_dsm}
\alias{surface_dsm}
\title{Digital Surface Model}
\usage{
surface_dsm(
  in_cat,
  dem_rts = NULL,
  out_rts,
  alg,
  res = 0.25,
  z_min = 0,
  z_max = 80,
  LAS_select = "xyzcr",
  LAS_classes = NULL,
  ...
)
}
\description{
Generate a DSM. If \code{dem_rts} is provided, it will generate a normalized DSM (nDSM)
}
