% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{tile_poly2}
\alias{tile_poly2}
\title{Tile polygons}
\usage{
tile_poly2(
  in_gpkg,
  seg_poly_rsds,
  seg_id = "polyID",
  tile_names = NULL,
  overwrite = FALSE
)
}
\description{
Convert a single polygonal dataset (must be in GPKG format) to tiles
}
