% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{training_data}
\alias{training_data}
\title{Training Data (constructor)}
\usage{
training_data(name, dir, proj = getOption("misterRS.crs"), overwrite = FALSE)
}
\description{
Training Data (constructor)
}
