% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_VerticalImage.R
\name{VerticalImage}
\alias{VerticalImage}
\title{Vertical image}
\usage{
VerticalImage(inputs, out_RSDS, tileNames = NULL, overwrite = FALSE)
}
\description{
Composite image of two ortho PCA bands and a normalized digital surface model
}
