% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts.R
\name{vts}
\alias{vts}
\title{Vector tileset}
\description{
Vector tileset.
}
\section{Super class}{
\code{\link[misterRS:xts]{misterRS::xts}} -> \code{vts}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{con}}{to database.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-vts-new}{\code{vts$new()}}
\item \href{#method-vts-finalize}{\code{vts$finalize()}}
\item \href{#method-vts-print}{\code{vts$print()}}
\item \href{#method-vts-temp_con}{\code{vts$temp_con()}}
\item \href{#method-vts-connect}{\code{vts$connect()}}
\item \href{#method-vts-connected}{\code{vts$connected()}}
\item \href{#method-vts-disconnect}{\code{vts$disconnect()}}
\item \href{#method-vts-transact}{\code{vts$transact()}}
\item \href{#method-vts-has_tiles}{\code{vts$has_tiles()}}
\item \href{#method-vts-fields}{\code{vts$fields()}}
\item \href{#method-vts-read_tile}{\code{vts$read_tile()}}
\item \href{#method-vts-read_from_pts}{\code{vts$read_from_pts()}}
\item \href{#method-vts-read_from_polys}{\code{vts$read_from_polys()}}
\item \href{#method-vts-append_geom}{\code{vts$append_geom()}}
\item \href{#method-vts-update_data}{\code{vts$update_data()}}
\item \href{#method-vts-add_attribute}{\code{vts$add_attribute()}}
\item \href{#method-vts-add_field}{\code{vts$add_field()}}
\item \href{#method-vts-index}{\code{vts$index()}}
\item \href{#method-vts-complete}{\code{vts$complete()}}
\item \href{#method-vts-clone}{\code{vts$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="misterRS" data-topic="xts" data-id="metadata"><a href='../../misterRS/html/xts.html#method-xts-metadata'><code>misterRS::xts$metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="misterRS" data-topic="xts" data-id="metadata_calc"><a href='../../misterRS/html/xts.html#method-xts-metadata_calc'><code>misterRS::xts$metadata_calc()</code></a></span></li>
<li><span class="pkg-link" data-pkg="misterRS" data-topic="xts" data-id="metadata_path"><a href='../../misterRS/html/xts.html#method-xts-metadata_path'><code>misterRS::xts$metadata_path()</code></a></span></li>
<li><span class="pkg-link" data-pkg="misterRS" data-topic="xts" data-id="metadata_read"><a href='../../misterRS/html/xts.html#method-xts-metadata_read'><code>misterRS::xts$metadata_read()</code></a></span></li>
<li><span class="pkg-link" data-pkg="misterRS" data-topic="xts" data-id="metadata_save"><a href='../../misterRS/html/xts.html#method-xts-metadata_save'><code>misterRS::xts$metadata_save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-new"></a>}}
\if{latex}{\out{\hypertarget{method-vts-new}{}}}
\subsection{Method \code{new()}}{
Initialize VTS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$new(
  id,
  name,
  dir,
  geom_type,
  geom_layer = "layer",
  id_field = "poly_id",
  crs = getOption("misterRS.crs"),
  ts = getOption("misterRS.ts")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{ID of VTS.}

\item{\code{name}}{Name of VTS.}

\item{\code{dir}}{Path to directory in which files will be created.}

\item{\code{geom_type}}{Type of geometry (ex.: POLYGON, POINT).}

\item{\code{geom_layer}}{Name of layer for geometry.}

\item{\code{crs}}{EPSG number.}

\item{\code{ts}}{Tileset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-vts-finalize}{}}}
\subsection{Method \code{finalize()}}{
Finalize.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-print"></a>}}
\if{latex}{\out{\hypertarget{method-vts-print}{}}}
\subsection{Method \code{print()}}{
Print.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-temp_con"></a>}}
\if{latex}{\out{\hypertarget{method-vts-temp_con}{}}}
\subsection{Method \code{temp_con()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$temp_con()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-connect"></a>}}
\if{latex}{\out{\hypertarget{method-vts-connect}{}}}
\subsection{Method \code{connect()}}{
Connect to Geopackage DB.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$connect(mod_spatialite = getOption("misterRS.mod_spatialite"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod_spatialite}}{Path to Spatialite directory.
temp_con = function(){
con <- DBI::dbConnect(RSQLite::SQLite(), dbname = self$gpkg)
withr::defer({DBI::dbDisconnect(con)}, envir = parent.frame(1))
return(con)
},}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-connected"></a>}}
\if{latex}{\out{\hypertarget{method-vts-connected}{}}}
\subsection{Method \code{connected()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$connected()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-vts-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect from Gepackage DB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-transact"></a>}}
\if{latex}{\out{\hypertarget{method-vts-transact}{}}}
\subsection{Method \code{transact()}}{
Execute SQL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$transact(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{Code for transaction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-has_tiles"></a>}}
\if{latex}{\out{\hypertarget{method-vts-has_tiles}{}}}
\subsection{Method \code{has_tiles()}}{
Check if has tiles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$has_tiles(tile_names, attribute)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tile_names}}{Name of tile.}

\item{\code{attribute}}{Name of attribute.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-fields"></a>}}
\if{latex}{\out{\hypertarget{method-vts-fields}{}}}
\subsection{Method \code{fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-read_tile"></a>}}
\if{latex}{\out{\hypertarget{method-vts-read_tile}{}}}
\subsection{Method \code{read_tile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$read_tile(tile_name, fields = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-read_from_pts"></a>}}
\if{latex}{\out{\hypertarget{method-vts-read_from_pts}{}}}
\subsection{Method \code{read_from_pts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$read_from_pts(pts, fields = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-read_from_polys"></a>}}
\if{latex}{\out{\hypertarget{method-vts-read_from_polys}{}}}
\subsection{Method \code{read_from_polys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$read_from_polys(polys, fields = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-append_geom"></a>}}
\if{latex}{\out{\hypertarget{method-vts-append_geom}{}}}
\subsection{Method \code{append_geom()}}{
Write tile.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$append_geom(data, tile_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{sf object being written.}

\item{\code{tile_name}}{Name of tile.}

\item{\code{attribute}}{Name of attribute.}

\item{\code{overwrite}}{Overwrite this tile.

NOTE: I tried overwriting geometry using this method and it seems to have somehow corrupted the GPKG.
The st_write function would execute without error, but the geometry would be invalid and wouldn't show up when loaded into QGIS
This needs to be fixed, but will require additional investimgation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-update_data"></a>}}
\if{latex}{\out{\hypertarget{method-vts-update_data}{}}}
\subsection{Method \code{update_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$update_data(data, tile_name, attribute, overwrite = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-add_attribute"></a>}}
\if{latex}{\out{\hypertarget{method-vts-add_attribute}{}}}
\subsection{Method \code{add_attribute()}}{
Add new attribute to tile registry.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$add_attribute(attribute)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attribute}}{Name of attribute being added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-add_field"></a>}}
\if{latex}{\out{\hypertarget{method-vts-add_field}{}}}
\subsection{Method \code{add_field()}}{
Add new field to geometry layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$add_field(field_name, field_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field_name}}{Name of field being added.}

\item{\code{field_type}}{Data type of field.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-index"></a>}}
\if{latex}{\out{\hypertarget{method-vts-index}{}}}
\subsection{Method \code{index()}}{
Create a database index on a column.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$index(field_name = "tile_name")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field_name}}{Name of field being indexed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-complete"></a>}}
\if{latex}{\out{\hypertarget{method-vts-complete}{}}}
\subsection{Method \code{complete()}}{
Is this VTS complete?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$complete(attribute, tile_names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attribute}}{Name of attribute.}

\item{\code{tile_names}}{Name of tiles. If set to NULL, all tiles will be selected.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-vts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{vts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
