% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vts.R
\name{vts_clear_attribute_set}
\alias{vts_clear_attribute_set}
\title{Clear tileset reg attributes}
\usage{
vts_clear_attribute_set(in_vts, attribute_set, tile_name = NULL)
}
\description{
Clear tileset reg attributes
}
