% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vts.R
\name{vts_row_count}
\alias{vts_row_count}
\title{Count numbers of rows in vectir tileset per tile}
\usage{
vts_row_count(in_vts, out_file, overwrite = FALSE)
}
\description{
Count numbers of rows in vectir tileset per tile
}
