% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_VWF.R
\name{VWF}
\alias{VWF}
\title{Variable window filter}
\usage{
VWF(
  CHM_RSDS,
  ttops_RSDS,
  winFun,
  minHeight,
  tileNames = NULL,
  clusters = 1,
  overwrite = FALSE
)
}
\description{
Variable window filter
}
