% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Segments.R
\name{WatershedSegmentation}
\alias{WatershedSegmentation}
\title{Watershed Segmentation}
\usage{
WatershedSegmentation(
  out_RSDS,
  CHM_RSDS,
  ttops_RSDS,
  minCrownHgt = 0.2,
  tileNames = NULL,
  overwrite = FALSE
)
}
\description{
Watershed Segmentation
}
