% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_xts.R
\name{xts}
\alias{xts}
\title{Generic tileset}
\description{
Generic tileset.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-xts-new}{\code{xts$new()}}
\item \href{#method-xts-clone}{\code{xts$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xts-new"></a>}}
\if{latex}{\out{\hypertarget{method-xts-new}{}}}
\subsection{Method \code{new()}}{
Initialize XTS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xts$new(id, name, dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{ID of XTS.}

\item{\code{name}}{Name of XTS.}

\item{\code{dir}}{Path to directory in which files will be created.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-xts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
