% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_misc.R
\name{ZipChunk}
\alias{ZipChunk}
\title{Zip a RSDS by chunks}
\usage{
ZipChunk(
  RSDS,
  deliveryDir,
  fileprefix,
  chunkSize = 100,
  zipExe = "C:/RBuildTools/3.5/bin/zip.exe"
)
}
\description{
Zip a RSDS by chunks
}
