% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_mask.R
\name{canopy_mask}
\alias{canopy_mask}
\title{Create canopy mask}
\usage{
canopy_mask(
  seg_class_ras_rsds,
  out_rsds,
  canopy_classes,
  canopy_edits = NULL,
  openings = 1,
  opening_radius = 0.5,
  tile_names = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{canopy_classes}{names of raster classes that are included in the canopy}

\item{openings}{the number of times that a morphological opening will be applied}

\item{opening_radius}{radius of morphological opening window}
}
\description{
Create canopy mask
}
