% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_seg_poly}
\alias{classify_seg_poly}
\title{#' Create classifier
#'
#' @export}
\usage{
classify_seg_poly(
  classifier_file,
  seg_poly_vts,
  seg_class_poly_vts,
  class_edits,
  metrics,
  seg_id,
  ...
)
}
\description{

}
\details{

}
