% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_seg_poly}
\alias{classify_seg_poly}
\title{Classify polygonal segments}
\usage{
classify_seg_poly(
  classifier_file,
  seg_vts,
  iteration,
  class_table,
  class_edits,
  ...
)
}
\description{
Classify polygonal segments
}
