% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_seg_ras}
\alias{classify_seg_ras}
\title{Classify raster segments}
\usage{
classify_seg_ras(seg_vts, seg_rts, seg_class_rts, iteration, seg_id, ...)
}
\description{
This needs \code{classify_seg_poly} to be run first
}
