% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.exe_tile_worker}
\alias{.exe_tile_worker}
\title{Run a worker in parallel or serial}
\usage{
.exe_tile_worker(
  tile_names,
  worker,
  clusters = getOption("misterRS.clusters"),
  cluster_eval = NULL
)
}
\arguments{
\item{cluster_eval}{extra code that gets evaluated on each cluster}
}
\description{
Run a worker in parallel or serial
}
