% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Surfaces.R
\name{MakeNDSM}
\alias{MakeNDSM}
\title{Canopy Height Model}
\usage{
MakeNDSM(
  LAS_RSDS,
  DEM_RSDS,
  out_RSDS,
  nDSMres,
  zMin,
  zMax,
  maxEdge = c(0, 1),
  subCircle = 0,
  thresholds = c(0, 2, 5, 10, 15, 20, 25, 30, 35, 40),
  LASselect = "xyzcr",
  LASclasses = NULL,
  tileNames = NULL,
  clusters = 1,
  overwrite = FALSE
)
}
\description{
Canopy Height Model
}
