% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_vts}
\alias{merge_vts}
\title{Merge multiple VTS into a single one according to zones}
\usage{
merge_vts(in_vts_list, out_vts, zones_path, zone_field, id_field = NULL, ...)
}
\description{
Merge multiple VTS into a single one according to zones
}
