% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Mosaic.R
\name{Mosaic}
\alias{Mosaic}
\title{Mosaic a RS Dataset}
\usage{
Mosaic(
  RSDS,
  overlap = "nbuffs",
  shp_clip = FALSE,
  outFile = NULL,
  overwrite = FALSE,
  tileNames = NULL
)
}
\description{
Mosaic a RS Dataset
}
