% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_metrics.R
\name{seg_metrics_las}
\alias{seg_metrics_las}
\title{Segment Metrics - LAS}
\usage{
seg_metrics_las(
  seg_ras_rsds,
  seg_poly_rsds,
  in_cat,
  dem_rsds,
  out_rsds,
  metric_fun,
  z_min,
  z_max,
  seg_id,
  prefix = NULL,
  tile_names = NULL,
  overwrite = FALSE
)
}
\description{
Segment Metrics - LAS
}
