% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_metrics.R
\name{seg_metrics_las}
\alias{seg_metrics_las}
\title{Segment Metrics - LAS}
\usage{
seg_metrics_las(
  seg_rts,
  seg_vts,
  in_cat,
  dem_rts,
  attribute_set,
  z_min = 0,
  z_max = 100,
  ground_class = NULL,
  full_class = NULL,
  rgb = NULL,
  ...
)
}
\description{
Segment Metrics - LAS
}
