% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_metrics.R
\name{seg_metrics_spec}
\alias{seg_metrics_spec}
\title{Segment Metrics - Spectral}
\usage{
seg_metrics_spec(
  seg_ras_rsds,
  seg_poly_rsds,
  img_rsds,
  out_rsds,
  bands = c(R = 1, G = 2, B = 3),
  zonalFun = c("mean", "sd"),
  seg_id,
  tile_names = NULL,
  overwrite = FALSE,
  prefix = NULL
)
}
\description{
Segment Metrics - Spectral
}
