% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_metrics.R
\name{seg_metrics_tex}
\alias{seg_metrics_tex}
\title{Segment Metrics - Textural}
\usage{
seg_metrics_tex(
  seg_ras_rsds,
  seg_poly_rsds,
  img_rsds,
  out_rsds,
  seg_id,
  band = 1,
  tile_names = NULL,
  overwrite = FALSE,
  prefix = "",
  n_grey = 16
)
}
\description{
Segment Metrics - Textural
}
