% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_SegMetrics.R
\name{SegMetricsTex}
\alias{SegMetricsTex}
\title{Segment Metrics - Textural}
\usage{
SegMetricsTex(
  segRas_RSDS,
  segPoly_RSDS,
  PCA_RSDS,
  nDSM_RSDS,
  out_RSDS,
  segID,
  tileNames = NULL,
  overwrite = FALSE
)
}
\description{
Segment Metrics - Textural
}
