% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaces.R
\name{surface_dsm}
\alias{surface_dsm}
\title{Digital Surface Model}
\usage{
surface_dsm(
  in_cat,
  dem_rsds = NULL,
  out_rsds,
  res = 0.25,
  z_min = 0,
  z_max = 80,
  max_edge = c(0, 1),
  subcircle = 0,
  thresholds = c(0, 2, 5, 10, 15, 20, 25, 30, 35, 40),
  LAS_select = "xyzcr",
  LAS_classes = NULL,
  tile_names = NULL,
  overwrite = FALSE
)
}
\description{
Generate a DSM. If \code{dem_rsds} is provided, it will generate a normalized DSM (nDSM)
}
