% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{training_data_extract}
\alias{training_data_extract}
\title{Extract Training data}
\usage{
training_data_extract(training_data, attribute_set, seg_vts, overwrite = FALSE)
}
\description{
Extract Training data
}
