% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{training_data_extract}
\alias{training_data_extract}
\title{Extract Training data}
\usage{
training_data_extract(
  training_pts,
  seg_poly_rsds,
  metrics,
  seg_id,
  overwrite = FALSE
)
}
\description{
Extract Training data
}
