% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_Segments.R
\name{WatershedSegmentation}
\alias{WatershedSegmentation}
\title{Watershed Segmentation}
\usage{
WatershedSegmentation(
  out_RSDS,
  CHM_RSDS,
  ttops_RSDS,
  OSGeoPath = "C:/OSGeo4W64",
  tileNames = NULL,
  overwrite = FALSE
)
}
\description{
Watershed Segmentation
}
