% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts.R
\name{xts}
\alias{xts}
\title{Generic tileset}
\description{
Generic tileset.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-xts-new}{\code{xts$new()}}
\item \href{#method-xts-metadata}{\code{xts$metadata()}}
\item \href{#method-xts-metadata_path}{\code{xts$metadata_path()}}
\item \href{#method-xts-metadata_read}{\code{xts$metadata_read()}}
\item \href{#method-xts-metadata_calc}{\code{xts$metadata_calc()}}
\item \href{#method-xts-metadata_save}{\code{xts$metadata_save()}}
\item \href{#method-xts-clone}{\code{xts$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xts-new"></a>}}
\if{latex}{\out{\hypertarget{method-xts-new}{}}}
\subsection{Method \code{new()}}{
Initialize XTS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xts$new(id, name, dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{ID of XTS.}

\item{\code{name}}{Name of XTS.}

\item{\code{dir}}{Path to directory in which files will be created.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xts-metadata"></a>}}
\if{latex}{\out{\hypertarget{method-xts-metadata}{}}}
\subsection{Method \code{metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xts$metadata(attribute)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xts-metadata_path"></a>}}
\if{latex}{\out{\hypertarget{method-xts-metadata_path}{}}}
\subsection{Method \code{metadata_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xts$metadata_path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xts-metadata_read"></a>}}
\if{latex}{\out{\hypertarget{method-xts-metadata_read}{}}}
\subsection{Method \code{metadata_read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xts$metadata_read()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xts-metadata_calc"></a>}}
\if{latex}{\out{\hypertarget{method-xts-metadata_calc}{}}}
\subsection{Method \code{metadata_calc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xts$metadata_calc(attribute)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xts-metadata_save"></a>}}
\if{latex}{\out{\hypertarget{method-xts-metadata_save}{}}}
\subsection{Method \code{metadata_save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xts$metadata_save(metadata)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-xts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{xts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
